Added a new state checking method isFinished() to Scavenger. This method returns a boolean which when true indicates that the Scavenger will be unable to load any new results in the future (because the Scraper can provide no more images for analysis). It checks that both image and result buffers are empty and the tasks tasks which refil those buffers are finished before returning true to inidcate that the Scacenger is finished. It was added to allow clients to use Scavenger safely without any chance of Scavenger blocking. Clients can now do: while (!scavenger.isFinished()) { 	while (scavenger.hasNextResult()) { 		scavenger.loadNextImage(); 	} } scavenger.exit();
To enable this, Future<?> instances are used by Scavenger to store the status of imageDataBufferTask and huntingTask. These instances are named imageBufferStatus and huntingStatus respectively. huntingTask now accepts imageBufferStatus. This enables huntingTask to know when imageDataBufferTask is finished and no longer refilling the image buffer. This allows us to avoid hanging on imageBuffer.take(). We still use take() because under normal conditions we want to block until an image is available. If however the buffer is empty and we know the buffer is no longer being refilled, we do not want to block, so instead we finish up the Task, which results in huntingStatus being marked as finished (huntingStatus.isDone() == true).