Fixed a long standing issue where Tesseract would print a warning to STDOUT whenever an image whose metadata contains a dpi of 0 or whose metadata did not contain a dpi value. This could be triggered by every image if the image source has previously stripped all image metadata. To fix this, a default dpi of 300 is set in the Tesseract instance. This value is now used whenever such images are encountered. No error message is displayed. A dpi of 300 was chosen since the accuracy with a high dpi setting analysing a low dpi image should be better than setting a low dpi then analysing a high dpi image (since small text may be ignored). The README was updated to remove the reference to this known defect since it is now fixed.
Modified OCRTess4J to utilise the Tesseract1 implementation rather than the Tesseract implementation. Tesseract1 uses JNA to utilise native methods for increased performance. OCR is used extensively by ScreenShotScavenger so it was decided it would be worthwhile to use the fastest implementation. This change resulted in a minor performance improvement, with no affect on the accuracy of the results.

Tidied up the output of ExampleUsage. It now also prints the reason an image was flagged in live time (as opposed to only when the results were printed after a call to printResultsAndExit().