Fixed issue where PrntscScraper could potentially block indefinitely on a call to nextImage() if images failed to download. PrntscScraper's nextImage() method now has a timeout of 10 seconds to prevent indefinite blocking. If the timeout is reached and we don't have an image, we throw a NoImageAvailableException(). Fixed the handling of interrupts in nextImage(). We now check the interrupted status of the thread before doing anything. If the thread is interrupted we shutdown and throw a NoImageAvailableException. The same approach is used to handle InterruptedExceptions thrown by the BlockingQueue imageBuffer.
This should help particularly  if cloudlflare block your access (by ip) to prntsc servers and images are failing to download. Previously, this could result in indefinite blocking by nextImage() as we used BlockingQueue's take() method. If images are all failing to download due to ip blocking, the take() method will never return as we won't have any new images to add to the queue. The consumer will be starved.
Now we timeout instead, which at least returns control to the client.

Also:
removed blocking from PrntscScraper's shutdown method. No need to block while waiting for resources to be released.

Fixed shutdown() in PrntscScraper. It was calling pool.shutdown(), which executes already submitted tasks, which is not the behavious we want since the images downloaded in these tasks will not be used and it greatly increases the time it takes for resources to be released when shutdown() is called. Instead pool.shutdownNow() is called which causes all remaining tasks to be discarded.